topic模块结束

+++ src/main/java/com/example/topic/controller/TopicController.java
@@ -3,14 +3,13 @@ package com.example.topic.controller;
 import com.example.topic.domain.Log;
 import com.example.topic.domain.TopicPo;
 import com.example.topic.feign.LogServiceApi;
+import com.example.topic.feign.PicturesServiceApi;
 import com.example.topic.service.TopicService;
-import com.example.topic.util.FileUtils;
 import com.example.topic.util.ResponseUtil;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -20,17 +19,20 @@ import java.util.List;
 /**
  * @author 对象模型标准组
  * @time 2019/12/9
+ *
+ * @author 宋澳龙
+ * @time 2019/12/15
  */
 @RestController
-@RequestMapping("/topicService")
 public class TopicController {
 
     @Autowired
     private TopicService topicService;
     @Autowired
     private LogServiceApi logServiceApi;
-    @Value("${web.path}")
-    private String path;
+    @Autowired
+    private PicturesServiceApi picturesServiceApi;
+
 
     /**
      * 管理员获取专题列表
@@ -40,19 +42,24 @@ public class TopicController {
      * @return 专题列表
      */
     @ApiOperation("管理员获取专题列表 /adminList")
-    @GetMapping("/admin/topics")
+    @GetMapping(value = "/admin/topics",produces = "application/json; charset=utf-8" )
     public Object adminList(@RequestParam(defaultValue = "1") Integer page,
                        @RequestParam(defaultValue = "10") Integer limit){
         Object retObj;
         Log log = new Log();
         log.setType(0);
         log.setActions("管理员获取专题列表");
-        PageHelper.startPage(page,limit);
-        List<TopicPo> topicList = topicService.list();
-        PageInfo<TopicPo> topicPageInfo = new PageInfo<>(topicList);
-        List<TopicPo> pageList = topicPageInfo.getList();
-        retObj = ResponseUtil.ok(pageList);
-        log.setStatusCode(1);
+        if(page<0||limit<0){
+            retObj = ResponseUtil.paramNotAllowed();
+            log.setStatusCode(0);
+        }else{
+            PageHelper.startPage(page,limit);
+            List<TopicPo> topicList = topicService.list();
+            PageInfo<TopicPo> topicPageInfo = new PageInfo<>(topicList);
+            List<TopicPo> pageList = topicPageInfo.getList();
+            retObj = ResponseUtil.ok(pageList);
+            log.setStatusCode(1);
+        }
         logServiceApi.addLog(log);
         return retObj;
     }
@@ -64,19 +71,24 @@ public class TopicController {
      * @return 专题详情
      */
     @ApiOperation("管理员获取专题详情 /detail")
-    @GetMapping("/admin/topics/{id}")
+    @GetMapping(value = "/admin/topics/{id}",produces = "application/json; charset=utf-8")
     public Object adminDetail(@NotNull @PathVariable(value = "id") Integer id) {
         Object retObj;
         Log log = new Log();
         log.setType(0);
         log.setActions("管理员获取专题详情");
-        TopicPo detail = topicService.detail(id);
-        if(detail==null){
-            retObj = ResponseUtil.fail(650,"该话题是无效话题");
+        if(id<0){
+            retObj = ResponseUtil.paramNotAllowed();
             log.setStatusCode(0);
         }else{
-            retObj = ResponseUtil.ok(detail);
-            log.setStatusCode(1);
+            TopicPo detail = topicService.detail(id);
+            if(detail==null){
+                retObj = ResponseUtil.fail(650,"该话题是无效话题");
+                log.setStatusCode(0);
+            }else{
+                retObj = ResponseUtil.ok(detail);
+                log.setStatusCode(1);
+            }
         }
         logServiceApi.addLog(log);
         return retObj;
@@ -90,38 +102,21 @@ public class TopicController {
      * @return 专题列表
      */
     @ApiOperation("用户获取专题列表 /list")
-    @GetMapping("/topics")
+    @GetMapping(value = "/topics",produces = "application/json; charset=utf-8")
     public Object list(@RequestParam(defaultValue = "1") Integer page,
                             @RequestParam(defaultValue = "10") Integer limit){
         Object retObj;
-        PageHelper.startPage(page,limit);
-        List<TopicPo> topicList = topicService.list();
-        PageInfo<TopicPo> topicPageInfo = new PageInfo<>(topicList);
-        List<TopicPo> pageList = topicPageInfo.getList();
-        retObj = ResponseUtil.ok(pageList);
+        if(page<0||limit<0){
+            retObj = ResponseUtil.paramNotAllowed();
+        }else {
+            PageHelper.startPage(page,limit);
+            List<TopicPo> topicList = topicService.list();
+            PageInfo<TopicPo> topicPageInfo = new PageInfo<>(topicList);
+            List<TopicPo> pageList = topicPageInfo.getList();
+            retObj = ResponseUtil.ok(pageList);
+        }
         return retObj;
     }
-
-//    /**
-//     * 上传图片到服务器 返回图片地址
-//     *
-//     * @param file
-//     * @return url 图片在服务器上的地址
-//     */
-//    @ApiOperation("上传图片")
-//    @PostMapping("/pic")
-//    public Object pic(@RequestParam("file") MultipartFile file){
-//        String localPath = "/picture";
-//        String fileName = file.getOriginalFilename();
-//        String newFile = FileUtils.upload(file,localPath,fileName);
-//        Object retObj;
-//        if(newFile != null){
-//            retObj = ResponseUtil.ok(localPath+"\\"+newFile);
-//        }else{
-//            retObj = ResponseUtil.fail(505,"图片上传失败");
-//        }
-//        return retObj;
-//    }
 
     /**
      * 专题详情
@@ -129,31 +124,52 @@ public class TopicController {
      * @param id 专题ID
      * @return 专题详情
      */
-    @ApiOperation("用户获取专题详情 /detail")
-    @GetMapping("/topics/{id}")
+    @ApiOperation( "用户获取专题详情 /detail")
+    @GetMapping(value ="/topics/{id}",produces = "application/json; charset=utf-8")
     public Object detail(@NotNull @PathVariable(value = "id") Integer id) {
         Object retObj;
-        TopicPo detail = topicService.detail(id);
-        if(detail==null){
-            retObj = ResponseUtil.fail(650,"该话题是无效话题");
-        }else{
-            retObj = ResponseUtil.ok(detail);
+        if(id<0){
+            retObj = ResponseUtil.paramNotAllowed();
+        }else {
+            TopicPo detail = topicService.detail(id);
+            if(detail==null){
+                retObj = ResponseUtil.fail(650,"该话题是无效话题");
+            }else{
+                retObj = ResponseUtil.ok(detail);
+            }
         }
         return retObj;
     }
+
+    /**
+     * 上传图片到服务器 返回图片地址
+     *
+     * @param image
+     * @return url 图片在服务器上的地址
+     */
+    @ApiOperation("上传图片")
+    @PostMapping("/picture")
+    public Object postPicture(@RequestParam("image") MultipartFile image){
+        Object retObj = picturesServiceApi.postPicture(image);
+        return  retObj;
+    }
+
 
     @ApiOperation("管理员添加专题 /create")
     /**
      @RequiresPermissions("admin:topic:create")
      @RequiresPermissionsDesc(menu = {"推广管理", "专题管理"}, button = "添加")
      */
-    @PostMapping("/topics")
+    @PostMapping(value = "/topics",produces = "application/json; charset=utf-8")
     public Object create(@RequestBody TopicPo topicPo) {
         Object retObj;
         Log log = new Log();
         log.setType(1);
         log.setActions("管理员添加专题");
-        if(topicService.beExist(topicPo)){
+        if(topicPo.getContent()==null&&topicPo.getPicUrlList()==null){
+            retObj = ResponseUtil.fail(652,"话题添加失败");
+            log.setStatusCode(0);
+        }else if(topicService.beExist(topicPo)){
             TopicPo detail = topicService.create(topicPo);
             if(detail.getId()==-1){
                 retObj = ResponseUtil.fail(652,"话题添加失败");
@@ -175,23 +191,27 @@ public class TopicController {
      @RequiresPermissions("admin:topic:update")
      @RequiresPermissionsDesc(menu = {"推广管理", "专题管理"}, button = "编辑")
      */
-    @PutMapping("/topics/{id}")
+    @PutMapping(value = "/topics/{id}",produces = "application/json; charset=utf-8")
     public Object update(@PathVariable Integer id,@RequestBody TopicPo topicPo) {
         Object retObj;
         Log log = new Log();
         log.setType(2);
         log.setActions("管理员编辑专题");
-        if(topicService.detail(id)==null){
-            retObj = ResponseUtil.fail(650,"该话题是无效话题");
-            log.setStatusCode(0);
-        }else{
-            TopicPo detail = topicService.update(topicPo,id);
-            if(detail.getId()==-1){
-                retObj = ResponseUtil.fail(651,"话题更新失败");
+        if(id<0){
+            retObj = ResponseUtil.paramNotAllowed();
+        }else {
+            if(topicService.detail(id)==null){
+                retObj = ResponseUtil.fail(650,"该话题是无效话题");
                 log.setStatusCode(0);
-            }else {
-                retObj = ResponseUtil.ok(detail);
-                log.setStatusCode(1);
+            }else{
+                TopicPo detail = topicService.update(topicPo,id);
+                if(detail.getId()==-1||detail.getPicUrlList()==null||detail.getContent()==null){
+                    retObj = ResponseUtil.fail(651,"话题更新失败");
+                    log.setStatusCode(0);
+                }else {
+                    retObj = ResponseUtil.ok(detail);
+                    log.setStatusCode(1);
+                }
             }
         }
         logServiceApi.addLog(log);
@@ -209,16 +229,20 @@ public class TopicController {
         Log log = new Log();
         log.setType(3);
         log.setActions("管理员删除专题");
-        if(topicService.detail(id)==null){
-            retObj = ResponseUtil.fail(650,"该话题是无效话题");
-            log.setStatusCode(0);
-        }else{
-            if(topicService.delete(id)==1){
-                retObj = ResponseUtil.ok();
-                log.setStatusCode(1);
-            }else{
-                retObj = ResponseUtil.fail(653,"话题删除失败");
+        if(id<0){
+            retObj = ResponseUtil.paramNotAllowed();
+        }else {
+            if(topicService.detail(id)==null){
+                retObj = ResponseUtil.fail(650,"该话题是无效话题");
                 log.setStatusCode(0);
+            }else{
+                if(topicService.delete(id)==1){
+                    retObj = ResponseUtil.ok();
+                    log.setStatusCode(1);
+                }else{
+                    retObj = ResponseUtil.fail(653,"话题删除失败");
+                    log.setStatusCode(0);
+                }
             }
         }
         logServiceApi.addLog(log);
